requirements.txt
# --- Core Dependencies ---
torch>=2.1.0             # Base deep learning library (check sd-mecha requirement)
safetensors>=0.4.2       # For efficient model loading/saving (check sd-mecha requirement)
numpy>=1.26.0            # Numerical operations (check sd-mecha requirement)
PyYAML>=6.0.1            # For reading YAML configs (check sd-mecha requirement)
omegaconf                # For advanced configuration management (used by Hydra)
hydra-core               # Core Hydra library for configuration

# --- sd-mecha Dependency ---
# sd-mecha==0.0.27       # OLD VERSION - COMMENTED OUT
sd-mecha                 # Use the latest compatible version, or specify one explicitly like >=0.1.0 if known

# --- Optimization Libraries ---
bayesian-optimization    # Keep if you want to retain BayesOpt as an option
optuna                   # ADDED: For Optuna optimizer support
# hyperopt                 # Keep if you plan to add Hyperopt later

# --- Image/Scoring/CLIP Related ---
Pillow                   # Image processing
clip                     # Original OpenAI CLIP (might be pulled by others)
open_clip_torch          # Alternative CLIP implementation (often preferred)
transformers             # Hugging Face library (used by many scorers like BLIP, CLIP)
# openai-clip            # Usually redundant if 'clip' or 'open_clip_torch' is present
# aesthetic-predictor-v2-5 # Pulled via pip install aesthetic-predictor-v2-5 directly
hpsv2                    # Pulled via pip install hpsv2 directly
# Other potential scorer dependencies might be pulled automatically by the scorer libraries themselves

# --- Plotting & Visualization ---
matplotlib               # Basic plotting (used by Optuna/Artist)
seaborn                  # Statistical data visualization (might be used by Artist)
plotly                   # Interactive plots (used by Optuna visualizations)
kaleido                  # Needed to save Plotly figures as static images

# --- sd-mecha Specific Dependencies (from its requirements.txt) ---
click>=8.1.7             # Used by sd-mecha scripts (good to have for compatibility)
scipy>=1.14.0            # Scientific computing (check sd-mecha requirement)
tqdm>=4.66.3             # Progress bars (check sd-mecha requirement)
fuzzywuzzy>=0.18.0       # For fuzzy string matching (used by sd-mecha resolve)
python-Levenshtein       # Speeds up fuzzywuzzy (optional but recommended)

# --- Other Potential Dependencies ---
lightgbm                 # Seems unrelated unless used by a specific scorer? Review if needed.
scikit-learn             # Used by some ML tasks, possibly by a scorer? Review if needed.
tensordict               # Appears related to PyTorch/distributed? Review if needed by models/scorers.
timm                     # Image models library, possibly used by a scorer? Review if needed.
fairscale                # Distributed training/sharding library. Unlikely needed unless doing specific advanced things.
# pynput                   # REMOVED: Was for old hotkey system, we'll use asyncio/other methods

# --- Testing ---
# pytest                   # Keep if you have tests or plan to add them